Set up game logic files.
Set up a gameboard containing 6, 10, 20 cards depending on the round.

Set up a factory function to track player picks via a set.

Set up a function to shuffle the cards of the gameboard.

Set up a function to test if the player failed the game when they make a choice.

Currently setting up a React component to pick a card.